assets/src/client.ts

export default class SRPClient {
  private getPrime(): bigint {
    return this.N;
  }

  private getGenerator(): bigint {
    return this.g;
  }

  private getKey(): bigint {
    return this.k;
  }

  private hash(...args: bigint[]): bigint {
    // 实现 hash 函数，这里需要根据你的具体需求来实现
    // 例如，可以使用 SHA-256 或其他哈希算法
    // 注意：BigInt 和哈希算法的结合可能需要特别处理
    return BigInt(/* ... */);
  }

  public generatePasswordHash(salt: bigint, identity: string, password: string): bigint {
    return this.hash(
      salt,
      this.hash(BigInt(Buffer.from(identity + ':' + password).toString('hex')))
    );
  }

  public generatePreMasterSecret(a: bigint, B: bigint, x: bigint, u: bigint): bigint {
    const N = this.getPrime();
    const g = this.getGenerator();
    const k = this.getKey();

    let B2 = B - (k * (g ** x % N));

    if (B2 < 0n) {
      B2 = N - BigInt(Math.abs(Number(B2)));
      B2 = B2 % N;
    }

    return (B2 ** (a + (u * x))) % N;
  }

  public generateVerifier(x: bigint): bigint {
    return (this.getGenerator() ** x) % this.getPrime();
  }

  public generatePublic(privateKey: bigint): bigint {
    return (this.getGenerator() ** privateKey) % this.getPrime();
  }
}
